# Make fast blur performance reasonable