Closes #2511 

It’s possible to push it even further using a ring buffer and fixed point arithmetic, but that would at least double the amount of code, and this version already gives a nice performance boost.

<details>
<summary>Benchmark</summary>
fast blur: sigma 3.0    time:   [10.750 ms 10.758 ms 10.766 ms]
                        change: [-52.185% -52.000% -51.832%] (p = 0.00 < 0.05)

fast blur: sigma 7.0    time:   [10.851 ms 10.873 ms 10.896 ms]
                        change: [-51.702% -51.598% -51.494%] (p = 0.00 < 0.05)

fast blur: sigma 50.0   time:   [11.371 ms 11.426 ms 11.486 ms]
                        change: [-50.958% -50.681% -50.399%] (p = 0.00 < 0.05)

gaussian blur: sigma 3.0
                        time:   [9.4690 ms 9.7881 ms 10.273 ms]

gaussian blur: sigma 7.0
                        time:   [20.815 ms 20.922 ms 21.042 ms]

gaussian blur: sigma 50.0
                        time:   [173.84 ms 174.37 ms 174.92 ms]

gaussian blur (dynamic image): sigma 3.0
                        time:   [5.2525 ms 5.2611 ms 5.2702 ms]

gaussian blur (dynamic image): sigma 7.0
                        time:   [13.056 ms 13.084 ms 13.118 ms]

gaussian blur (dynamic image): sigma 50.0
                        time:   [94.538 ms 94.824 ms 95.138 ms]
</details>