@kornelski 2025-07-03
It looks okay, but I'm not capable of verifying it is correct from just looking at the code. Last time I implemented such thing, I've had a bunch of off-by-ones or mixed-up offsets when radius was greater than image size.

The current tests only check pixels of an easy case, and the rest just checks for panics.

Maybe you could add more tests?