@awxkee 2025-07-03
Done. 
I believe it's not truly feasible to track all errors mentally in such algorithms, so I fuzz such algorithms, just as I fuzzed this one.