@awxkee 2025-07-06
I’ve added a few asserts to make sure everything will go according to plan.